--- wasm/examples/open_url/dev.flowingspdg.wasm.sdPlugin/index.html
@@ -1,17 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <title>Go(WASM)</title>
-    <link rel="stylesheet" type="text/css" href="./sdpi.css">
-</head>
-
-<body>
-    <!-- wasm_exec.jsを読み込む -->
-    <script src="./wasm_exec.js"></script>
-    <script src="./std.js"></script>
-    <button onclick="$SD.openUrl('https://www.elgato.com/')">CLICK TO OPEN</button>
-</body>
-
-</html>
